requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
async-timeout==4.0.3
attrs==24.2.0
bcrypt==4.2.0
blinker==1.6.2
cachetools==5.5.0
certifi==2023.5.7
cffi==1.17.1
charset-normalizer==3.2.0
click==8.1.5
coloredlogs==15.0.1
cryptography==43.0.3
deprecation==2.1.0
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.0
fastapi-cli==0.0.5
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
flatbuffers==23.5.26
frozenlist==1.4.1
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.65.0
gotrue==2.9.2
greenlet==2.0.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.6
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.2
humanfriendly==10.0
hyperframe==6.0.1
idna==3.4
imageio==2.31.1
itsdangerous==2.1.2
Jinja2==3.1.2
jose==1.0.0
lazy_loader==0.3
llvmlite==0.40.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
networkx==3.1
numba==0.57.1
numpy==1.24.4
oauthlib==3.2.2
onnxruntime==1.15.1
opencv-python-headless==4.8.0.74
packaging==23.1
passlib==1.7.4
Pillow==10.0.0
platformdirs==3.9.1
pooch==1.7.0
postgrest==0.17.1
propcache==0.2.0
proto-plus==1.24.0
protobuf==4.23.4
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PyMatting==1.1.8
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.12
PyWavelets==1.4.1
PyYAML==6.0.2
realtime==2.0.5
rembg==2.0.50
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.9.2
rsa==4.9
scikit-image==0.21.0
scipy==1.11.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.19
starlette==0.38.6
storage3==0.8.1
StrEnum==0.4.15
stripe==9.12.0
supabase==2.9.0
supafunc==0.6.1
sympy==1.12
tifffile==2023.7.10
tqdm==4.65.0
typer==0.12.5
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.0.3
uvicorn==0.31.0
uvloop==0.20.0
watchfiles==0.24.0
websockets==13.1
Werkzeug==2.3.6
yarl==1.14.0
